Poetry Blog Analysis/requirements.txt

streamlit
requests
python-dotenv
sentence-transformers
weaviate-client
scikit-learn
nltk
rouge
bert-score

